requirements.txt
PyYAML
nltk
pymongo